Refactor LOGIN_DAL and LOAIHANG_DAL classes
This commit refactors the LOGIN_DAL and LOAIHANG_DAL classes in the DAO namespace. The LOGIN_DAL class has been moved from the DAL folder to the DAO folder, and the LOAIHANG_DAL class has been deleted from the DAO folder and replaced with an updated version. These changes improve the organization and structure of the code.